站点签到可扩展框架
(cherry picked from commit 27d0725dc06c64eaaa70957fb6d0f3f0ddd46013)
